// src/app/model/articolo.ts
export interface Articolo {

    idArticolo?: number;
    Titolo?: string;
    Abstract?: string;
    DOI?: string;
    ISSN?: string;
    ISBN?: string;
    AnnoPubblicazione?: number;
    idConferenzaPresentazione?: number;
    Nome?: string;
    Data?: string;
    Luogo?: string;
    idGiornale?: number;
    g_titolo?: string;
    num_art?: number;
    
  }